refactor(ui): category options and handle category changes moved to blogController
